PR Comments aangepast.
